+++ src/contentScript/messenger/registration.ts
@@ -32,11 +32,9 @@ import {
   reactivateTab,
   removePersistedExtension,
 } from "@/contentScript/lifecycle"; // 196 strictNullCheck errors
-import { runBrick } from "@/contentScript/executor"; // Depends on background/messenger to pass strictNullCheck
 import { clearDynamicElements } from "@/contentScript/pageEditor/dynamic/clearDynamicElements"; // Depends on contentScript/lifecycle to pass strictNullCheck
 import { runStarterBrickReader } from "@/contentScript/pageEditor/dynamic/runStarterBrickReader"; // Depends on contentScript/messenger to pass strictNullCheck
 import { updateDynamicElement } from "@/contentScript/pageEditor/dynamic/updateDynamicElement"; // Depends on contentScript/lifecycle to pass strictNullCheck
-import { runBlockPreview } from "@/contentScript/pageEditor/runBlockPreview"; // Depends on background/messenger
 import { resetTab } from "@/contentScript/pageEditor/resetTab"; // Depends on contentScript/lifecycle to pass strictNullCheck
 import { runRendererBlock } from "@/contentScript/pageEditor/runRendererBlock"; // Depends on background/messenger
 import { runHeadlessPipeline } from "@/contentScript/pipelineProtocol/runHeadlessPipeline"; // Depends on background/messenger
@@ -52,7 +50,6 @@ declare global {
     RESET_TAB: typeof resetTab;
     ACTIVATE_PRERENDERED_TAB: typeof activatePrerenderedTab;
 
-    RUN_SINGLE_BLOCK: typeof runBlockPreview;
     RUN_RENDERER_BLOCK: typeof runRendererBlock;
 
     CLEAR_DYNAMIC_ELEMENTS: typeof clearDynamicElements;
@@ -61,8 +58,6 @@ declare global {
 
     INSTALLED_EXTENSION_POINTS: typeof getActiveExtensionPoints;
     ENSURE_EXTENSION_POINTS_INSTALLED: typeof ensureInstalled;
-
-    RUN_BRICK: typeof runBrick;
 
     RUN_RENDERER_PIPELINE: typeof runRendererPipeline;
     RUN_HEADLESS_PIPELINE: typeof runHeadlessPipeline;
@@ -77,7 +72,6 @@ export default function registerMessenger(): void {
     RESET_TAB: resetTab,
     ACTIVATE_PRERENDERED_TAB: activatePrerenderedTab,
 
-    RUN_SINGLE_BLOCK: runBlockPreview,
     RUN_RENDERER_BLOCK: runRendererBlock,
 
     CLEAR_DYNAMIC_ELEMENTS: clearDynamicElements,
@@ -86,8 +80,6 @@ export default function registerMessenger(): void {
 
     INSTALLED_EXTENSION_POINTS: getActiveExtensionPoints,
     ENSURE_EXTENSION_POINTS_INSTALLED: ensureInstalled,
-
-    RUN_BRICK: runBrick,
 
     RUN_RENDERER_PIPELINE: runRendererPipeline,
     RUN_HEADLESS_PIPELINE: runHeadlessPipeline,
