+++ src/contentScript/messenger/strict/api.ts
@@ -70,3 +70,5 @@ export const runMapArgs = getMethod("RUN_MAP_ARGS");
 export const getCopilotHostData = getMethod("GET_COPILOT_HOST_DATA");
 
 export const showLoginBanner = getMethod("SHOW_LOGIN_BANNER");
+export const runBlock = getMethod("RUN_SINGLE_BLOCK");
+export const runBrick = getMethod("RUN_BRICK");
