+++ src/contentScript/messenger/strict/registration.ts
@@ -58,6 +58,8 @@ import {
 import { runMapArgs } from "@/contentScript/pipelineProtocol/runMapArgs";
 import { getCopilotHostData } from "@/contrib/automationanywhere/SetCopilotDataEffect";
 import { showBannerFromConfig } from "@/contentScript/integrations/deferredLoginController";
+import { runBlockPreview } from "@/contentScript/pageEditor/runBlockPreview";
+import { runBrick } from "@/contentScript/executor";
 
 declare global {
   interface MessengerMethods {
@@ -97,6 +99,8 @@ declare global {
     RUN_MAP_ARGS: typeof runMapArgs;
     GET_COPILOT_HOST_DATA: typeof getCopilotHostData;
     SHOW_LOGIN_BANNER: typeof showBannerFromConfig;
+    RUN_SINGLE_BLOCK: typeof runBlockPreview;
+    RUN_BRICK: typeof runBrick;
   }
 }
 export default function registerMessenger(): void {
@@ -137,5 +141,7 @@ export default function registerMessenger(): void {
     RUN_MAP_ARGS: runMapArgs,
     GET_COPILOT_HOST_DATA: getCopilotHostData,
     SHOW_LOGIN_BANNER: showBannerFromConfig,
+    RUN_SINGLE_BLOCK: runBlockPreview,
+    RUN_BRICK: runBrick,
   });
 }
