+++ src/telemetry/trace.ts
@@ -29,6 +29,7 @@ import {
   type RenderedArgs,
 } from "@/types/runtimeTypes";
 import { deleteDatabase } from "@/utils/idbUtils";
+import { type Nullishable } from "@/utils/nullishUtils";
 
 const DATABASE_NAME = "TRACE";
 const ENTRY_OBJECT_STORE = "traces";
@@ -40,12 +41,12 @@ export type TraceRecordMeta = {
    *
    * `null` for ad-hoc block execution.
    */
-  extensionId: UUID | null;
+  extensionId: Nullishable<UUID>;
 
   /**
    * Extension run id. Unique run id to correlate trace elements from the same extension run.
    */
-  runId: UUID;
+  runId: Nullishable<UUID>;
 
   /**
    * Branches to the brick execution
@@ -61,7 +62,7 @@ export type TraceRecordMeta = {
   /**
    * Unique id to identify the block in the Page Editor across runs.
    */
-  blockInstanceId: UUID;
+  blockInstanceId: Nullishable<UUID>;
 
   /**
    * The registry id of the block.
